Jsons and cmake files renewed, packageDownloader.download_single_package() fixed (lambda wanted a vector that was never given)
