+++ main-server/CMakeLists.txt
@@ -15,7 +15,6 @@ include_directories(
         ${MONGOCXX_INCLUDE_DIR}
         ${BSONCXX_INCLUDE_DIR}
         include
-        ../common-include
 )
 
 #Add path to lib
@@ -28,8 +27,6 @@ link_directories(
 find_package(cpprestsdk REQUIRED)
 find_package(folly REQUIRED)
 find_package(GFlags REQUIRED)
-# find_package(GTest REQUIRED)
-# find_package(GMock REQUIRED)
 
 
 add_library(kumys_lib STATIC
@@ -42,8 +39,6 @@ add_library(kumys_lib STATIC
         include/RequestHandler.h
         include/DatabaseManager.h
         include/PackageDownloader.h
-        ../common-include/LightJson.h
-        ../common-include/HeavyJson.h
 )
 
 
@@ -60,5 +55,5 @@ add_executable(kumys_server src/main.cpp)
 target_link_libraries(kumys_server PRIVATE kumys_lib)
 
 #Include tests
-# enable_testing()
-# add_subdirectory(tests)
+enable_testing()
+add_subdirectory(tests)
