+++ main-server/tests/CMakeLists.txt
@@ -1,3 +1,9 @@
+include(FetchContent)
+FetchContent_Declare(
+        googletest
+        URL https://github.com/google/googletest/archive/03597a01ee50ed33e9dfd640b249b4be3799d395.zip
+)
+
 add_executable(tests
         test_main.cpp
         test_httpserver.cpp
@@ -13,8 +19,7 @@ target_link_libraries(tests
         mongocxx
         bsoncxx
         Folly::folly
-        GTest::GTest
-        GMock::GMock
+        gtest gmock gtest_main
 )
 
 include(GoogleTest)
